design: Chip 디자인 수정
- 임시 Close 아이콘 제거
- 컴포넌트 디자인 일부 적용
