--- src/common/components/Chip.tsx
@@ -1,37 +1,0 @@
-import { Icon } from '@chakra-ui/react';
-import * as S from '../styles/Chip.styles';
-
-interface ChipProps {
-  label: string;
-  closable?: boolean;
-  onClose?: () => void;
-}
-
-function Chip({ label, closable, onClose }: ChipProps) {
-  return (
-    <S.Chip>
-      <span>{label}</span>
-      {closable && (
-        <S.CloseButton onClick={onClose}>
-          {/* NOTE: 임시 아이콘 */}
-          <Icon name="close">
-            <svg
-              width="16"
-              height="16"
-              viewBox="0 0 16 16"
-              fill="none"
-              xmlns="http://www.w3.org/2000/svg"
-            >
-              <path
-                d="M8.00052 8.93337L4.73385 12.2C4.61163 12.3223 4.45608 12.3834 4.26719 12.3834C4.0783 12.3834 3.92274 12.3223 3.80052 12.2C3.6783 12.0778 3.61719 11.9223 3.61719 11.7334C3.61719 11.5445 3.6783 11.3889 3.80052 11.2667L7.06719 8.00003L3.80052 4.73337C3.6783 4.61114 3.61719 4.45559 3.61719 4.2667C3.61719 4.07781 3.6783 3.92225 3.80052 3.80003C3.92274 3.67781 4.0783 3.6167 4.26719 3.6167C4.45608 3.6167 4.61163 3.67781 4.73385 3.80003L8.00052 7.0667L11.2672 3.80003C11.3894 3.67781 11.545 3.6167 11.7339 3.6167C11.9227 3.6167 12.0783 3.67781 12.2005 3.80003C12.3227 3.92225 12.3839 4.07781 12.3839 4.2667C12.3839 4.45559 12.3227 4.61114 12.2005 4.73337L8.93385 8.00003L12.2005 11.2667C12.3227 11.3889 12.3839 11.5445 12.3839 11.7334C12.3839 11.9223 12.3227 12.0778 12.2005 12.2C12.0783 12.3223 11.9227 12.3834 11.7339 12.3834C11.545 12.3834 11.3894 12.3223 11.2672 12.2L8.00052 8.93337Z"
-                fill="#1A1A1A"
-              />
-            </svg>
-          </Icon>
-        </S.CloseButton>
-      )}
-    </S.Chip>
-  );
-}
-
-export default Chip;
