--- src/common/styles/Chip.styles.ts
@@ -1,17 +1,0 @@
-import styled from 'styled-components';
-
-export const Chip = styled.span`
-  display: inline-flex;
-  align-items: center;
-  padding: 0.5rem;
-  border-radius: 0.75rem;
-  background-color: #e6e6e6;
-`;
-
-export const CloseButton = styled.button`
-  display: inline-flex;
-  justify-content: center;
-  cursor: pointer;
-  align-items: center;
-  margin-left: 2px;
-`;
